+++ src/plugins/gta3/std.data/data_traits/ide.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -35,10 +35,8 @@ using fx5_9e = tuple<int, real_t, real_t, real_t, real_t, ipair,  ipair, ipair, 
 //
 struct ide_traits : public data_traits
 {
-    static const bool can_cache         = true;     // Can this store get cached?
-    static const bool is_reversed_kv    = false;    // Does the key contains the data instead of the value in the key-value pair?
     static const bool has_sections      = true;     // Does this data file contains sections?
-    static const bool per_line_section  = false;    // Is the sections of this data file different on each line?
+    static const bool per_line_section  = false;
 
     // Detouring traits
     struct dtraits : modloader::dtraits::OpenFile
@@ -49,9 +47,6 @@ struct ide_traits : public data_traits
     
     // Detouring type
     using detour_type = modloader::OpenFileDetour<0x5B8428, dtraits>;
-
-    // Dominance Flags
-    using domflags_fn = datalib::domflags_fn<flag_RemoveIfNotExistInOneCustomButInDefault>;
 
     // Section slices
     using path_type = data_slice<>;