+++ src/plugins/gta3/std.data/data_traits/object.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -8,32 +8,22 @@
 using namespace modloader;
 using std::string;
 
-//
 struct object_traits : public data_traits
 {
-    static const bool can_cache         = true;     // Can this store get cached?
-    static const bool is_reversed_kv    = false;    // Does the key contains the data instead of the value in the key-value pair?
-    static const bool has_sections      = false;    // Does this data file contains sections?
-    static const bool per_line_section  = false;    // Is the sections of this data file different on each line?
+    static const bool has_sections      = false;
+    static const bool per_line_section  = false;
 
-    // End of file string
     static const bool has_eof_string = true;
     static const char* eof_string() { return "*"; }
 
-    // Detouring traits
     struct dtraits : modloader::dtraits::OpenFile
     {
         static const char* what()       { return "object data"; }
         static const char* datafile()   { return "object.dat"; }
     };
     
-    // Detouring type
     using detour_type = modloader::OpenFileDetour<0x5B5444, dtraits>;
 
-    // Dominance Flags
-    using domflags_fn = datalib::domflags_fn<flag_RemoveIfNotExistInOneCustomButInDefault>;
-
-    // Object Data
     using key_type      = std::size_t;
     using value_type    = data_slice<modelname,
                                 real_t, real_t, real_t, real_t, real_t, real_t, real_t, int, int, int, int, int, delimopt,
@@ -52,13 +42,11 @@ public: // eof_string related
     { archive(this->eof); }
 };
 
-//
 using object_store = gta3::data_store<object_traits, std::map<
                         object_traits::key_type, object_traits::value_type
                         >>;
 
 
-// Objects Data Merger
 static auto xinit = initializer([](DataPlugin* plugin_ptr)
 {
     // XXX a perfect refresh needs to set all the CBaseModelInfo::m_wObjectInfoIndex to -1 before reloading the data file