+++ src/plugins/gta3/std.data/data_traits/plants.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -7,27 +7,18 @@
 #include "../data.hpp"
 using namespace modloader;
 
-//
 struct plants_traits : public data_traits
 {
-    static const bool can_cache         = true;     // Can this store get cached?
-    static const bool is_reversed_kv    = false;    // Does the key contains the data instead of the value in the key-value pair?
-    static const bool has_sections      = false;    // Does this data file contains sections?
-    static const bool per_line_section  = false;    // Is the sections of this data file different on each line?
+    static const bool has_sections      = false;
+    static const bool per_line_section  = false;
 
-    // Detouring traits
     struct dtraits : modloader::dtraits::OpenFile
     {
         static const char* what() { return "plants surface properties"; }
     };
     
-    // Detouring type
     using detour_type = modloader::OpenFileDetour<0x5DD3D1, dtraits>;
 
-    // Dominance Flags
-    using domflags_fn = datalib::domflags_fn<flag_RemoveIfNotExistInOneCustomButInDefault>;
-
-    // Plants data
     using key_type      = std::pair<std::size_t, int>;
     using value_type    = data_slice<std::string,
                             int, int, int, int, int, int, int, int, int, int,
@@ -39,13 +30,10 @@ struct plants_traits : public data_traits
     }
 };
 
-//
 using plants_store = gta3::data_store<plants_traits, std::map<
                         plants_traits::key_type, plants_traits::value_type
                         >>;
 
-
-// Plants Surface Properties Merger
 static auto xinit = initializer([](DataPlugin* plugin_ptr)
 {
     auto ReloadPlantsDat = [plugin_ptr]