+++ src/plugins/gta3/std.data/data_traits/surface.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -7,25 +7,17 @@
 #include "../data.hpp"
 using namespace modloader;
 
-//
 struct surface_traits : public data_traits
 {
-    static const bool can_cache         = true;     // Can this store get cached?
-    static const bool is_reversed_kv    = false;    // Does the key contains the data instead of the value in the key-value pair?
-    static const bool has_sections      = false;    // Does this data file contains sections?
-    static const bool per_line_section  = false;    // Is the sections of this data file different on each line?
+    static const bool has_sections      = false;
+    static const bool per_line_section  = false;
 
-    // Detouring traits
     struct dtraits : modloader::dtraits::OpenFile
     {
         static const char* what() { return "surface adhesion limits"; }
     };
     
-    // Detouring type
     using detour_type = modloader::OpenFileDetour<0x55D100, dtraits>;
-
-    // Dominance Flags
-    using domflags_fn = datalib::domflags_fn<flag_RemoveIfNotExistInOneCustomButInDefault>;
 
     // The first string from the line doesn't matter, it's just a helper to the human reader.
     // The following values are real numbers, '---' like values or nothing
@@ -45,13 +37,10 @@ public:
     { archive(this->adhesion_line); }
 };
 
-//
 using surface_store = gta3::data_store<surface_traits, std::map<
                         surface_traits::key_type, surface_traits::value_type
                         >>;
 
-
-//
 static auto xinit = initializer([](DataPlugin* plugin_ptr)
 {
     auto ReloadSurfaceInfo = std::bind(injector::thiscall<void(void*)>::call<0x55F420>, mem_ptr(0xB79538).get<void>());