+++ src/plugins/gta3/std.data/data_traits/tracks.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -9,7 +9,6 @@ using namespace modloader;
 
 struct tracks_traits
 {
-    // Detouring traits
     struct dtraits : modloader::dtraits::ReadAndInterpretTrackFile
     {
         static const char* what() { return "train track"; }
@@ -19,11 +18,9 @@ struct tracks_traits
 template<uintptr_t addr>
 using OpenTracksDetour = modloader::ReadAndInterpretTrackFileDetour<addr, tracks_traits::dtraits>;
 
-
-// Population Cycle Properties Detour
 static auto xinit = initializer([](DataPlugin* plugin_ptr)
 {
-    auto ReloadTracks = injector::cstd<void()>::call<0x6F7440>; // reinstalling this is kinda of dangerous I guess
+    //auto ReloadTracks = injector::cstd<void()>::call<0x6F7440>;
     plugin_ptr->AddDetour("tracks.dat", no_reinstall, OpenTracksDetour<0x6F7470>());
     plugin_ptr->AddDetour("tracks2.dat", no_reinstall, OpenTracksDetour<0x6F74BC>());
     plugin_ptr->AddDetour("tracks3.dat", no_reinstall, OpenTracksDetour<0x6F7496>());