+++ src/plugins/gta3/std.data/data_traits/weapon.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2015  LINK/2012 <dma_2012@hotmail.com>
+ * Copyright (C) 2015  LINK/2012 <dma_2012@hotmail.com>
  * Licensed under GNU GPL v3, see LICENSE at top level directory.
  * 
  */
@@ -11,8 +11,6 @@ using std::string;
 //
 struct weapon_traits : public data_traits
 {
-    static const bool can_cache         = true;     // Can this store get cached?
-    static const bool is_reversed_kv    = false;    // Does the key contains the data instead of the value in the key-value pair?
     static const bool has_sections      = true;     // Does this data file contains sections?
     static const bool per_line_section  = true;     // Is the sections of this data file different on each line?
 
@@ -29,10 +27,6 @@ struct weapon_traits : public data_traits
     
     // Detouring type
     using detour_type = modloader::OpenFileDetour<0x5BE68A, dtraits>;
-
-    // Dominance Flags
-    using domflags_fn = datalib::domflags_fn<flag_RemoveIfNotExistInOneCustomButInDefault>;
-
 
     // Section slices
     using melee_type = data_slice<char, insen<string>, string, real_t, real_t, int, int, int, string, int, hex<int>, string>;